requirements.txt
# ZERO Voice Assistant - Requirements
# Python 3.9+

# ============================================
# Audio - Wake Word Detection
# ============================================
pvporcupine==2.2.0

# ============================================
# Audio - Speech-to-Text (STT)
# ============================================
deepgram-sdk==3.2.0

# ============================================
# Audio - Text-to-Speech (TTS)
# ============================================
TTS==0.22.0

# ============================================
# Audio - I/O
# ============================================
pyaudio==0.2.14
sounddevice==0.4.6
pydub==0.25.1

# ============================================
# Natural Language Understanding (NLU)
# ============================================
spacy==3.7.2
# Note: After installation, run: python -m spacy download en_core_web_sm
openai==1.6.0
dateparser==1.2.0

# ============================================
# Skills - Weather
# ============================================
requests==2.31.0
pyowm==3.3.0

# ============================================
# Skills - App Control
# ============================================
psutil==5.9.6
pywin32==306; sys_platform == 'win32'
pyobjc-framework-Cocoa==10.1; sys_platform == 'darwin'
pyobjc-framework-ApplicationServices==10.1; sys_platform == 'darwin'

# ============================================
# Skills - Timer
# ============================================
playsound==1.3.0

# ============================================
# Configuration & Utilities
# ============================================
PyYAML==6.0.1
python-dotenv==1.0.0

# ============================================
# Frontend - CLI
# ============================================
rich==13.7.0
questionary==2.0.1

# ============================================
# Frontend - System Tray
# ============================================
pystray==0.19.5
Pillow==10.2.0

# ============================================
# Frontend - Notifications
# ============================================
plyer==2.1.0

# ============================================
# Testing & Development
# ============================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.0
pylint==3.0.3
mypy==1.8.0

# ============================================
# Packaging (for standalone executables)
# ============================================
pyinstaller==6.3.0
